Address PR review: deduplicate revert_to_stopped, update docs
- Move revert_to_stopped to minion_registry as a shared public utility,
  removing duplicate copies from attach.rs and resume.rs
- Update handle_attach doc comment to describe backend resolution and
  behavior when interactive attach isn't supported, replacing
  Claude-specific language
